first transaction from python

+++ contract.sol
@@ -4,20 +4,20 @@ contract WavesPresale {
     struct Sale
     {
         bytes32 txid;   
-        uint128 amount;
+        uint amount;
         uint date;
     }
     
     Sale[] public sales;
     uint32 public numberOfSales;
-    uint128 public totalTokens;
+    uint public totalTokens;
 
     function WavesPresale() {
         owner = msg.sender;
         numberOfSales = 0;
     }
 
-    function newSale(bytes32 txid, uint128 amount) {
+    function newSale(bytes32 txid, uint amount) {
         if (msg.sender != owner) return;
 
         sales.push(Sale({
@@ -28,5 +28,5 @@ contract WavesPresale {
         numberOfSales += 1;
         totalTokens += amount;
     }
-    
+
 }
